moves ObjectMapper class to configuration package.
