+++ core/src/main/java/org/arquillian/smart/testing/configuration/Configuration.java
@@ -15,7 +15,6 @@ import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import org.arquillian.smart.testing.Logger;
-import org.arquillian.smart.testing.ObjectMapper;
 import org.arquillian.smart.testing.RunMode;
 import org.arquillian.smart.testing.hub.storage.local.LocalStorage;
 import org.arquillian.smart.testing.hub.storage.local.LocalStorageFileAction;
